Push image tag transform code to plugin.
